Clean up idempotency tests
