--- tests/idempotency/BootstrapChecker.scala
@@ -1,5 +1,0 @@
-
-object Test {
-  def main(args: Array[String]): Unit =
-    IdempotencyCheck.checkIdempotency("out/tastyBootstrap/dotty1/dotty1", "out/tastyBootstrap/dotty2/dotty2")
-}
