--- tests/idempotency/CheckOrderIdempotency.scala
@@ -1,5 +1,0 @@
-
-object Test {
-  def main(args: Array[String]): Unit =
-    IdempotencyCheck.checkIdempotency("out/idempotency/orderIdempotency1", "out/idempotency/orderIdempotency2")
-}
