--- tests/idempotency/CheckPosIdempotency.scala
@@ -1,5 +1,0 @@
-
-object Test {
-  def main(args: Array[String]): Unit =
-    IdempotencyCheck.checkIdempotency("out/idempotency/posIdempotency1", "out/idempotency/posIdempotency2")
-}
