+++ tests/pos/order-idempotency-list/N.scala
